added in swagger and authentication

+++ Webservices/src/main/java/com/LuckBattle/Webservices/payload/response/JwtResponse.java
@@ -10,12 +10,12 @@ import java.util.List;
 public class JwtResponse {
     private String token;
     private String type = "Bearer";
-    private Long id;
+    private int id;
     private String username;
     private String email;
     private List<String> roles;
 
-    public JwtResponse(String token, Long id, String username, String email, List<String> roles) {
+    public JwtResponse(String token, int id, String username, String email, List<String> roles) {
         this.token = token;
         this.id = id;
         this.username = username;
